Added a check for v12
If we are using V12 we need to use ImageSharp2 instead of ImageSharp3. The powershell code checks if we our version starts with 12. And then we delete the Umbraco.Cms Package. Then we add the individual needed packages for umbraco. We add imageSharp2 Instead of 3 since we have experience load test spikes for v12 with ImageSharp3